PP-15510 rename setup service
