+++ src/main/java/uk/gov/pay/connector/gatewayaccount/service/AdyenAccountSetupService.java
@@ -3,22 +3,27 @@ package uk.gov.pay.connector.gatewayaccount.service;
 import com.google.inject.Inject;
 import com.google.inject.persist.Transactional;
 import uk.gov.pay.connector.gatewayaccount.dao.AdyenAccountSetupDao;
+import uk.gov.pay.connector.gatewayaccount.model.AdyenAccountSetupResponse;
 import uk.gov.pay.connector.gatewayaccount.model.AdyenAccountSetupStatus;
 import uk.gov.pay.connector.gatewayaccount.model.AdyenAccountSetupTask;
 import uk.gov.pay.connector.gatewayaccount.model.AdyenAccountSetupTaskEntity;
 import uk.gov.pay.connector.gatewayaccount.model.GatewayAccountEntity;
 
+import java.util.Arrays;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 import static uk.gov.pay.connector.gateway.PaymentGatewayName.ADYEN;
 
 public class AdyenAccountSetupService {
 
-    private final AdyenAccountSetupDao aydenAccountSetupDao;
+    private final AdyenAccountSetupDao adyenAccountSetupDao;
 
     @Inject
-    public AdyenAccountSetupService(AdyenAccountSetupDao aydenAccountSetupDao) {
-        this.aydenAccountSetupDao = aydenAccountSetupDao;
+    public AdyenAccountSetupService(AdyenAccountSetupDao adyenAccountSetupDao) {
+        this.adyenAccountSetupDao = adyenAccountSetupDao;
     }
 
     @Transactional
@@ -26,10 +31,43 @@ public class AdyenAccountSetupService {
         if (gatewayAccountEntity.isAdyenTestAccount()) {
             var gatewayAccountCredentialsEntity = gatewayAccountEntity.getRecentNonRetiredGatewayAccountCredentialsEntity(ADYEN.getName());
             List.of(AdyenAccountSetupTask.values()).forEach(task -> {
-                aydenAccountSetupDao.persist(new AdyenAccountSetupTaskEntity(gatewayAccountEntity, task, gatewayAccountCredentialsEntity, AdyenAccountSetupStatus.COMPLETED));
+                adyenAccountSetupDao.persist(new AdyenAccountSetupTaskEntity(gatewayAccountEntity, task, gatewayAccountCredentialsEntity, AdyenAccountSetupStatus.COMPLETED));
             });
         } else {
             throw new IllegalArgumentException("Gateway account type must be TEST and gateway name must be ADYEN");
         }
+    }
+
+    public AdyenAccountSetupResponse buildResponse(String serviceId, long gatewayAccountId, String credentialExternalId) {
+        AdyenAccountSetupResponse adyenAccountSetupResponse = new AdyenAccountSetupResponse();
+
+        adyenAccountSetupResponse.setServiceId(serviceId);
+        adyenAccountSetupResponse.setGatewayAccountId(gatewayAccountId);
+        adyenAccountSetupResponse.setCredentialExternalId(credentialExternalId);
+
+        adyenAccountSetupResponse.setTasks(getTasksWithStatus(gatewayAccountId));
+
+        return adyenAccountSetupResponse;
+    }
+
+    private HashMap<String, Map<String, AdyenAccountSetupStatus>> getTasksWithStatus(long gatewayAccountId) {
+        List<AdyenAccountSetupTaskEntity> taskEntities = adyenAccountSetupDao.findByGatewayAccountIdAndCredentialId(gatewayAccountId);
+
+        Map<String, AdyenAccountSetupStatus> taskStatusMap = taskEntities.stream()
+                .collect(Collectors.toMap(task -> task.getTask().getValue(), AdyenAccountSetupTaskEntity::getStatus));
+
+        var updatedTasks = new HashMap<String, Map<String, AdyenAccountSetupStatus>>();
+
+        Arrays.stream(AdyenAccountSetupTask.values()).forEach(task -> {
+            String taskName = task.getValue();
+
+            AdyenAccountSetupStatus status = taskStatusMap.getOrDefault(taskName, AdyenAccountSetupStatus.NOT_STARTED);
+
+            var statusHashMap = new HashMap<String, AdyenAccountSetupStatus>();
+
+            statusHashMap.put("status", status);
+            updatedTasks.put(taskName, statusHashMap);
+        });
+        return updatedTasks;
     }
 }
